refactor(os): absorb @routecraft/browser into @routecraft/os
Move the agentBrowser adapter and its test from @routecraft/browser into
@routecraft/os, the system-native host-capability package, and remove the
standalone browser package.

- Relocate src/adapters/agent-browser and the bun test into packages/os.
- Export agentBrowser from @routecraft/os; un-private the package and add
  agent-browser as an optional peer (lazy-loaded, install-time optional).
- Update the adapter's missing-peer error to name @routecraft/os.
- Rewrite the os README to document agentBrowser plus the migration note,
  and the planned shell/sandbox/host capabilities.
- Repoint size-limit, ci-cd.md, and docs-site references; drop the stale
  'os is private' note now that os publishes.
- Changeset: @routecraft/os minor; drop browser from the pending changeset.

agentBrowser's factory, options, and result shape are unchanged; consumers
update the import from @routecraft/browser to @routecraft/os.
