+++ .changeset/os-absorbs-browser.md
@@ -1,0 +1,5 @@
+---
+"@routecraft/os": minor
+---
+
+Fold browser automation into `@routecraft/os`. `agentBrowser()` now ships from `@routecraft/os` instead of the standalone `@routecraft/browser` package, which is deprecated. Update imports from `@routecraft/browser` to `@routecraft/os`; the factory, options, and result shape are unchanged.
