+++ .changeset/wild-llamas-share.md
@@ -1,6 +1,5 @@
 ---
 "@routecraft/ai": minor
-"@routecraft/browser": minor
 ---
 
 Declare core as a peer dependency with a real semver range (plus a workspace devDependency for development) instead of duplicating it as a regular dependency.
