--- packages/browser/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@routecraft/browser",
-  "version": "0.5.0",
-  "description": "Agent browser automation adapter for Routecraft",
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsup src/index.ts --format esm,cjs --dts",
-    "test": "vitest",
-    "prepublishOnly": "bun run build"
-  },
-  "devDependencies": {
-    "@routecraft/routecraft": "workspace:*",
-    "@routecraft/testing": "workspace:*",
-    "agent-browser": "^0.17.1",
-    "vitest": "^4.1.5"
-  },
-  "peerDependencies": {
-    "@routecraft/routecraft": ">=0.5.0 <1.0.0",
-    "agent-browser": "^0.17.1"
-  },
-  "peerDependenciesMeta": {
-    "agent-browser": {
-      "optional": true
-    }
-  },
-  "keywords": [
-    "routecraft",
-    "browser",
-    "automation",
-    "agent-browser",
-    "playwright"
-  ],
-  "author": "routecraftjs",
-  "license": "Apache-2.0",
-  "homepage": "https://routecraft.dev",
-  "bugs": "https://github.com/routecraftjs/routecraft/issues",
-  "repository": "https://github.com/routecraftjs/routecraft",
-  "publishConfig": {
-    "access": "public"
-  }
-}
