--- packages/browser/src/index.ts
@@ -1,7 +1,0 @@
-export {
-  agentBrowser,
-  type AgentBrowserBaseOptions,
-  type AgentBrowserCommandMap,
-  type AgentBrowserCommand,
-  type AgentBrowserResult,
-} from "./adapters/agent-browser/index.ts";
