+++ packages/os/src/index.ts
@@ -1,5 +1,12 @@
 // @routecraft/os -- system-native host capabilities.
-// Planned members: shell (sandboxed by default), sandbox, browser automation
-// (agentBrowser, merging in from @routecraft/browser), and future host
-// capabilities such as clipboard and notifications. Adapters land in future PRs.
-export {};
+// Current members: agentBrowser (browser automation). Planned: shell
+// (sandboxed by default), sandbox, and host primitives such as clipboard,
+// notifications, and process management. See .standards/package-boundaries.md.
+
+export {
+  agentBrowser,
+  type AgentBrowserBaseOptions,
+  type AgentBrowserCommandMap,
+  type AgentBrowserCommand,
+  type AgentBrowserResult,
+} from "./adapters/agent-browser/index.ts";
